Autoload UI scene and make better warp transition
Now instead of having to manually put in the UI viewport in each scene yourself, the project automatically adds it to the game tree.

And you also don't have to add in a ColorRect background for each scene so the CanvasModulate Global script modulates the color of everything, the new ColorRect "Fade" node in the UI is used instead.