Add failsafes and assert custom errors
Uhhhh so if you reset a warpScene path in a resource that had one set previously, the bug from before still happens where it opens the dialog box when there isn't any text??? Maybe it's an engine issue. Should open up one later or check if it's been reported already.

It's weird, it's literally an empty string, and yet the warpScene still evaluates to true?