Check if there's dialog before zooming to an image
The game will now only open the dialogbox if there is any dialog text to display. If there isn't, just open the zoomImage. When it's clicked, free it from the scene.

Now I just need to check if there's a warpScene, too, so you can chain all three options together.

Sigh, I really should be using a state machine for this...