Set custom mouse cursors once
And only change the default cursor shape afterwards. This allows for Control nodes to be able to use these custom cursors without needing code.

Also deleted the animation tracks for the player's collision shape since having the points be animated every frame fucks with setting the default cursor when the fps is uncapped.