Fix dialog text freezing when shorter  than 10
Figured out why the "magic number" worked. Just needed to scroll down a couple lines :P

The condition for skipping the typewriter effect and closing the box was hardcoded to be a value at least 11 characters, and "...\n\nOk." just so happened to be 8 characters. idk why that condition is even there tbh